+++ browser/src/components/CodeViewer/CodeDiffView/index.tsx
@@ -23,11 +23,13 @@ const useStyle = createStyles(
         height: 100%;
         display: flex;
         flex-direction: column;
-        ${maxHeight
-          ? css`
+        ${
+          maxHeight
+            ? css`
               max-height: ${maxHeight}px;
             `
-          : ''}
+            : ''
+        }
       `,
       editor: css`
         height: 100%;