+++ src/utils/ripgrep.ts
@@ -1,17 +1,25 @@
 import { execFile } from 'child_process';
+import createDebug from 'debug';
 import path from 'pathe';
 import { findActualExecutable } from 'spawn-rx';
-import { fileURLToPath, resolve } from 'url';
+import { fileURLToPath } from 'url';
+import { isLocal } from './isLocal';
+
+const debug = createDebug('neovate:utils:ripgrep');
 
 const __filename = fileURLToPath(import.meta.url);
-const __dirname = resolve(__filename, '../');
+const __dirname = path.dirname(__filename);
+
+const rootDir = isLocal()
+  ? path.resolve(__dirname, '../../')
+  : path.resolve(__dirname, '../');
 
 function ripgrepPath() {
   const { cmd } = findActualExecutable('rg', []);
   if (cmd !== 'rg') {
     return cmd;
   } else {
-    const rgRoot = path.resolve(__dirname, 'vendor', 'ripgrep');
+    const rgRoot = path.resolve(rootDir, 'vendor', 'ripgrep');
     if (process.platform === 'win32') {
       return path.resolve(rgRoot, 'x64-win32', 'rg.exe');
     } else {
@@ -35,6 +43,7 @@ export async function ripGrep(
       },
       (err, stdout) => {
         if (err) {
+          debug(`[Ripgrep] Error: ${err}`);
           resolve([]);
         } else {
           resolve(stdout.trim().split('\n').filter(Boolean));